refactor(login): split login modal in signIn and signUp modals
