+++ ui/design/src/components/SignInModal/eth-provider-modal-illustration.tsx
